\name{bootstrapSelectInput}
\alias{bootstrapSelectInput}
\title{bootstrap-select Input}
\usage{
  bootstrapSelectInput(inputId, label, choices,
    selected = NULL, multiple = FALSE, style = NULL,
    title = NULL, selectedTextFormat = NULL, header = NULL,
    icon = NULL, subtext = NULL, content = NULL,
    liveSearch = NULL, size = NULL, options = NULL)
}
\arguments{
  \item{inputId}{input id}

  \item{...}{see selectInput()}

  \item{options}{options passed as json object (not sure if
  it works yet)}
}
\description{
  ...
}

