\name{bootstrapSwitch}
\alias{bootstrapSwitch}
\title{Bootstrap switch}
\usage{
  bootstrapSwitch(inputId, ..., type = "make-switch",
    size = "switch-mini")
}
\arguments{
  \item{inputId}{input id}

  \item{...}{see checkboxInput()}

  \item{size}{size (e.g. 'switch-mini')}
}
\description{
  ...
}

