% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{create_element}
\alias{create_element}
\title{Create a virtual DOM element}
\usage{
create_element(tag, ...)
}
\arguments{
\item{tag}{Character string specifying the HTML tag name}

\item{...}{Named arguments become props, unnamed arguments become children}
}
\value{
A list representing a virtual DOM element
}
\description{
Create a virtual DOM element
}
\keyword{internal}
