% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{create_manual_packages_index}
\alias{create_manual_packages_index}
\title{Manually create PACKAGES index file}
\usage{
create_manual_packages_index(repo_dir, tgz_files)
}
\arguments{
\item{repo_dir}{Repository directory}

\item{tgz_files}{Vector of .tgz file paths}
}
\description{
Manually create PACKAGES index file
}
\keyword{internal}
