% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f10.cwas_2celltype_allmodel_cal.R
\name{cwas_2celltype_allmodel_cal}
\alias{cwas_2celltype_allmodel_cal}
\title{Calculate All Models with Two Cell Types as Variables}
\usage{
cwas_2celltype_allmodel_cal(
  cwas.data,
  selected_celltype = NULL,
  linkfunction = "binomial",
  interaction = F
)
}
\arguments{
\item{cwas.data}{The CWAS dataset containing the rates and phenotypes of cell types.}

\item{selected_celltype}{A character vector specifying the cell types to consider for modeling (default is NULL, which uses all unique cell types in the dataset).}

\item{linkfunction}{The link function to use for modeling (default is "binomial").}

\item{interaction}{Logical indicating whether to consider interaction effects between cell types (default is FALSE).}
}
\value{
A list of calculated models for each combination of two cell types.
}
\description{
This function computes all possible models involving combinations of two specified cell types from the CWAS data.
}
\examples{
cwas_2celltype_allmodel_cal(cwas.test.data)
}
