% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f9.cwas_addPower.R
\name{cwas_addPower}
\alias{cwas_addPower}
\title{Add Power Information to Fitted Models}
\usage{
cwas_addPower(
  mymodel,
  yournsim = 10,
  OR_cutoff_up = NULL,
  OR_cutoff_down = NULL,
  CCI = T,
  interaction = F,
  Tablename = "Tablename"
)
}
\arguments{
\item{mymodel}{A list of fitted models obtained from previous analyses.}

\item{yournsim}{Number of simulations to perform for power estimation (default is 10).}

\item{OR_cutoff_up}{Upper cutoff value for odds ratio (OR) (optional).}

\item{OR_cutoff_down}{Lower cutoff value for odds ratio (OR) (optional).}

\item{CCI}{Logical indicating whether to consider cell change index (CCI) (default is TRUE).}

\item{interaction}{Logical indicating whether to consider interaction effects (default is FALSE).}

\item{Tablename}{Name of the table to be created (default is "Tablename").}
}
\value{
A list of fitted models with added power and CI information.
}
\description{
This function adds power and confidence interval (CI) information to the given fitted models for analysis.
}
\examples{
cwas_addPower(cwas.test.data.single.model.best)

}
