% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f11.cwas_addSamples.R
\name{cwas_addSamples}
\alias{cwas_addSamples}
\title{Simulate Power Curve with Increasing Sample Size}
\usage{
cwas_addSamples(
  cwas.data.selected,
  selected_celltype,
  yournsim = 10,
  OR_cutoff_up = NULL,
  OR_cutoff_down = NULL,
  CCI = T,
  interaction = F,
  increased_sample_info = "rate"
)
}
\arguments{
\item{cwas.data.selected}{A list containing the selected models of CWAS analysis.}

\item{selected_celltype}{A character vector specifying the cell types of interest.}

\item{yournsim}{Number of simulations to perform for each sample size increment (default is 10).}

\item{OR_cutoff_up}{Upper cutoff value for odds ratio (OR) (optional).}

\item{OR_cutoff_down}{Lower cutoff value for odds ratio (OR) (optional).}

\item{CCI}{Logical indicating whether to consider cell change index (CCI) (default is TRUE).}

\item{interaction}{Logical indicating whether to consider interaction effects (default is FALSE).}

\item{increased_sample_info}{The method to increase sample size, such as "rate" or other specified factors.}
}
\value{
A simulated power curve plot for each specified cell type with varying sample sizes.
}
\description{
This function simulates a power curve by varying the sample size (X) and the power (Y) based on specified parameters.
}
