% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f8.cwas_autoselected_model.R
\name{cwas_autoselected_model}
\alias{cwas_autoselected_model}
\title{Automatically Select Best Model and Generate Forest Plot with Odds Ratios}
\usage{
cwas_autoselected_model(mdlist, method = c("Best"), interaction = F)
}
\arguments{
\item{mdlist}{A list of models obtained from previous analyses.}

\item{method}{Method for selecting the best model ("Best", "Equally Best", or "Integration") (default is "Best").
\itemize{
\item "Best": Selects the model with the lowest AIC (Akaike Information Criterion) value as the best model.
\item "Equally Best": Computes multiple models with equivalent AIC values based on KL (Kullback-Leibler) information, then selects the model with corrected positive results.
\item "Integration": Integrates all credible models to derive a combined inference.
}}

\item{interaction}{Logical indicating whether to consider interaction effects (default is FALSE).}
}
\value{
A list containing the selected best model and additional information.
}
\description{
This function automatically selects the best model from a list of models and generates a forest plot with odds ratios (ORs).
}
\examples{
 cwas_autoselected_model(cwas.test.data.single.model,method="Best")
}
