% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.cwas_build_model_data_v3.R
\name{cwas_build_model_data}
\alias{cwas_build_model_data}
\title{Prepare CWAS Data for Modeling}
\usage{
cwas_build_model_data(
  inputdata,
  Sample = "orig.ident",
  Phenotype,
  Celltype,
  Group = NULL,
  Subgroup = NULL,
  Covariate1 = NULL,
  Covariate2 = NULL,
  selected_celltype = NULL,
  Control_label = "Control",
  Disease_label = "Disease",
  Cellrate = NULL,
  genelist = NULL,
  geneset_score = NULL,
  add_therapeutic_targets = F
)
}
\arguments{
\item{inputdata}{A Seurat object or a metadata dataframe containing sample metadata, phenotype, cell type, and optional additional variables.}

\item{Sample}{The column name specifying sample identifiers.}

\item{Phenotype}{The column name specifying the phenotype variable.}

\item{Celltype}{The column name specifying the cell type variable.}

\item{Group}{(Optional) The column name specifying group information.}

\item{Subgroup}{(Optional) The column name specifying subgroup information.}

\item{Covariate1}{(Optional) The column name specifying the first covariate.}

\item{Covariate2}{(Optional) The column name specifying the second covariate.}

\item{selected_celltype}{(Optional) A vector of specific cell types to include in the analysis.}

\item{Control_label}{The label representing control samples (default is "Control").}

\item{Disease_label}{The label representing disease samples (default is "Disease").}

\item{Cellrate}{(Optional) The column name specifying precomputed cell rates (if available).}
}
\value{
A standardized dataframe suitable for CWAS analysis, containing sample identifiers, phenotype, group information, subgroup information, covariates, and cell rates.
}
\description{
This function prepares input data containing sample metadata, phenotype, cell type, and optional additional variables into a standard format suitable for subsequent analysis. The input can be a Seurat object or a metadata dataframe. It extracts necessary columns and computes cell rates if not provided.
}
\examples{
 cwas_build_model_data(inputdata = cwas.test,Sample = "GSM",Phenotype = "sample.condition",Celltype = "Subtype",Group = "GSE",Control_label="lymph.node",Disease_label="tumor",Cellrate = "Fraction")
}
