% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3.cwas_classic_comparision.R
\name{cwas_classic_comparision}
\alias{cwas_classic_comparision}
\title{Perform Classic Comparison of Cell Proportions and Phenotypes}
\usage{
cwas_classic_comparision(cwas.data)
}
\arguments{
\item{cwas.data}{The CWAS data containing information on cell types, phenotypes, and rates.}
}
\value{
A list containing the results of statistical tests and a list of plots for each cell type.
}
\description{
This function calculates classic comparisons of cell proportions and phenotypes based on the provided CWAS data.
}
\examples{
cwas_classic_comparision(cwas.test.data)
}
