% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.2_fill_missing_celltype_V2.R
\name{cwas_fill_missing_celltype}
\alias{cwas_fill_missing_celltype}
\title{Fill Missing Celltypes with Zero and Retain Sample Metadata}
\usage{
cwas_fill_missing_celltype(
  data,
  sample_col = "Sample",
  celltype_col = "Celltype"
)
}
\arguments{
\item{data}{A dataframe that contains at least sample, celltype, and rate columns.}

\item{sample_col}{A string specifying the column name for sample identifiers. Default is "Sample".}

\item{celltype_col}{A string specifying the column name for cell types. Default is "Celltype".}

\item{rate_col}{A string specifying the column name for the rate values. Default is "rate".}
}
\value{
A dataframe where missing sample-celltype combinations have been filled in with
a rate of 0, and all metadata related to the samples has been retained and matched.
}
\description{
This function fills missing cell type rates in the provided dataset by adding any missing
combinations of samples and cell types, setting their rate to 0. Additionally, it retains
and matches all other metadata related to the samples.
}
