% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f4.cwas_hausman_test.R
\name{cwas_hausman_test_all}
\alias{cwas_hausman_test_all}
\title{Conducts Hausman Test for each Celltype in  data}
\usage{
cwas_hausman_test_all(cwas.data, variable, family = "binomial")
}
\arguments{
\item{cwas.data}{Data frame containing  data}

\item{variable}{Name of the variable to test (dependent variable)}

\item{family}{Type of family for GLM (default is "binomial")}
}
\value{
A data frame with results of Hausman test for each Celltype
}
\description{
Conducts Hausman Test for each Celltype in  data
}
