% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f12.cwas_mediation_analysis.R
\name{cwas_mediation_analysis}
\alias{cwas_mediation_analysis}
\title{Conduct CWAS Mediation Analysis}
\usage{
cwas_mediation_analysis(
  cwas.data,
  Best_model,
  X.cell = NULL,
  X.gene = NULL,
  mediate.cell = NULL,
  mediate.gene = NULL,
  method = NULL,
  num_cores = NULL,
  fix_effect = NULL,
  random_effect = NULL
)
}
\arguments{
\item{cwas.data}{A data frame containing the CWAS results.}

\item{Best_model}{The best model selected for mediation analysis.}

\item{X.cell}{Vector specifying the cell types used as predictors in the mediation analysis. Default is NULL.}

\item{X.gene}{Vector specifying the genes used as predictors in the mediation analysis. Default is NULL.}

\item{mediate.cell}{Vector specifying the cell types to be used as mediators in the mediation analysis.}

\item{mediate.gene}{Vector specifying the genes to be used as mediators in the mediation analysis.}

\item{method}{The method for conducting mediation analysis. It should be one of "cell_mediate_cell", "gene_mediate_cell", or "cell_mediate_gene".}

\item{num_cores}{Number of CPU cores to be used for parallel computation. Default is NULL.}
}
\value{
The results of CWAS mediation analysis.
}
\description{
This function conducts CWAS (Cell-type Weighted Analysis of Single-cell data) mediation analysis based on the specified method.
}
