% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f5.2.cwas_model_cal_fix.R
\name{cwas_model_cal_fix}
\alias{cwas_model_cal_fix}
\title{CWAS Model Calculation for Fixed Effects}
\usage{
cwas_model_cal_fix(
  cwas.data,
  selectedcelltype = "selectedcelltype_name",
  link = "binomial"
)
}
\arguments{
\item{cwas.data}{A data frame containing the phenotype data and covariates.}

\item{selectedcelltype}{A character specifying the selected cell type for modeling.}

\item{link}{The link function to use in the model. Default is "binomial" for logistic regression.}
}
\value{
A list containing two elements:
\item{fitall}{A list of all fitted model objects that were successfully computed.}
\item{fitres_sorted}{A data frame containing summary statistics of the fitted models, sorted by AIC (Akaike Information Criterion).}
}
\description{
This function computes fixed effects models for phenotype data using generalized linear models (GLMs) with different link functions.
}
\details{
This function fits a series of fixed effects models using generalized linear models (GLMs) from the stats package. The models include combinations of fixed effects (rate, covariates) with different grouping structures (Subgroup, Group, Subgroup/Group). The function attempts to fit each model and captures successful results for further analysis.

The function outputs a list of fitted models (\code{fitall}) and a summary data frame (\code{fitres_sorted}) containing information about each model including model ID, cell type, formula, AIC, and p-value (based on the rate coefficient).
}
\examples{
cwas_model_cal_fix(cwas.test.data)
}
