% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f5.1.cwas_model_cal_mix.R
\name{cwas_model_cal_mix}
\alias{cwas_model_cal_mix}
\title{CWAS Model Calculation for Mixed Effects}
\usage{
cwas_model_cal_mix(
  cwas.data,
  selectedcelltype = "selectedcelltype",
  link = "binomial"
)
}
\arguments{
\item{cwas.data}{A data frame containing the phenotype data and covariates.}

\item{selectedcelltype}{A character specifying the selected cell type for modeling.}

\item{link}{The link function to use in the model. Default is "binomial" for logistic regression.}
}
\value{
A list containing two elements:
\item{fitall}{A list of all fitted model objects that were successfully computed.}
\item{fitres_sorted}{A data frame containing summary statistics of the fitted models, sorted by AIC (Akaike Information Criterion).}
}
\description{
This function computes mixed effects models for phenotype data using various combinations of fixed effects and random intercepts/slopes.
}
\details{
This function fits a series of mixed effects logistic regression models using the glmer function from the lme4 package. The models include combinations of fixed effects (rate, covariates) and random intercepts/slopes for different grouping structures (Subgroup, Group, Group/Subgroup). The function attempts to fit each model and captures successful results for further analysis.

The function outputs a list of fitted models (\code{fitall}) and a summary data frame (\code{fitres_sorted}) containing information about each model including model ID, cell type, formula, AIC, and p-value (based on the rate coefficient).
}
\examples{
cwas_model_cal_mix(cwas.test.data)

}
