% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7.cwas_model_heatmap.R
\name{cwas_model_heatmap}
\alias{cwas_model_heatmap}
\title{Plot Heatmap of AIC Values and P-values for All Models}
\usage{
cwas_model_heatmap(mdlist)
}
\arguments{
\item{mdlist}{A list of models containing AIC values and P-values.}
}
\description{
This function generates a heatmap displaying AIC (Akaike Information Criterion) values and P-values for all models.
}
\examples{
cwas_model_heatmap(cwas.test.data.single.model)

}
