% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f13.cwas_moderation_analysis.R
\name{cwas_moderation_analysis}
\alias{cwas_moderation_analysis}
\title{cwas_moderation_analysis}
\usage{
cwas_moderation_analysis(
  cwas.data,
  Best_model,
  X.cell = NULL,
  X.gene = NULL,
  Moderation.cell = NULL,
  Moderation.gene = NULL,
  method = NULL,
  num_cores = NULL
)
}
\arguments{
\item{cwas.data}{A data frame containing the input data for the analysis.}

\item{Best_model}{A model object containing information about the best model.}

\item{X.cell}{A string specifying the cell type for the moderation effect analysis. Default is NULL.}

\item{X.gene}{A string specifying the gene for the moderation effect analysis. Default is NULL.}

\item{Moderation.cell}{A string specifying the moderating cell type. Default is NULL.}

\item{Moderation.gene}{A string specifying the moderating gene. Default is NULL.}

\item{method}{A string specifying the method to use for the analysis. Acceptable values are "cell_moderation_cell", "gene_moderation_cell", or "cell_moderation_gene".}

\item{num_cores}{An integer specifying the number of cores to use for parallel computation. If NULL, sequential computation is performed. Default is NULL.}
}
\value{
An object containing the results of the moderation analysis.
}
\description{
This function performs moderation analysis between cells or genes. It calculates the moderation effects between cells, genes, or a combination of both based on the specified method.
}
\details{
\itemize{
\item If the \code{method} parameter is "cell_moderation_cell", the function performs cell-to-cell moderation effect analysis.
\item If the \code{method} parameter is "gene_moderation_cell", the function performs gene-to-cell moderation effect analysis. Depending on the \code{num_cores} parameter, it decides whether to perform sequential or parallel computation.
\item If the \code{method} parameter is "cell_moderation_gene", the function performs cell-to-gene moderation effect analysis. Depending on the \code{num_cores} parameter, it decides whether to perform sequential or parallel computation.
}
}
