% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2.cwas_pcadraw.R
\name{cwas_pcadraw}
\alias{cwas_pcadraw}
\title{Generate PCA Plot for CWAS Data}
\usage{
cwas_pcadraw(cwas.data, div_group = "Phenotype")
}
\arguments{
\item{cwas.data}{The CWAS data containing information on samples, cell types, rates, and the grouping variable.}

\item{div_group}{The name of the grouping variable used for color-coding in the PCA plot (default is "Phenotype").}
}
\value{
A PCA plot visualizing the CWAS data.
}
\description{
This function generates a PCA (Principal Component Analysis) plot for the given CWAS data based on cell type rates and a specified grouping variable.
}
\examples{
cwas_pcadraw(cwas.test.data)
}
