% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f6.cwas_plot_aic_pvalue.R
\name{cwas_plot_aic_pvalue}
\alias{cwas_plot_aic_pvalue}
\title{Plot AIC and P-value Scatter Plots for Each Cell Type}
\usage{
cwas_plot_aic_pvalue(mdlist, color = NULL)
}
\arguments{
\item{mdlist}{A list of models containing AIC values and P-values for each cell type.}

\item{color}{Optional. A vector of color codes to use for differentiating models in the scatter plots.}
}
\value{
A list of AIC and P-value Scatter Plots
}
\description{
This function generates separate scatter plots of AIC (Akaike Information Criterion) values against P-values for each cell type.
}
\examples{
cwas_plot_aic_pvalue(cwas.test.data.single.model)

}
