% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f4.cwas_regressiondraw.R
\name{cwas_regressiondraw}
\alias{cwas_regressiondraw}
\title{Perform Subset Regression Analysis for All Cell Types}
\usage{
cwas_regressiondraw(
  subcelldata,
  sampleID = c("Sample"),
  celltype = c("Celltype"),
  cellrate = c("rate"),
  div_group = c("Phenotype")
)
}
\arguments{
\item{subcelldata}{The subset of data containing sample IDs, cell types, cell rates, and phenotype groups.}

\item{sampleID}{The column name in \code{subcelldata} representing sample IDs (default is "Sample").}

\item{celltype}{The column name in \code{subcelldata} representing cell types (default is "Celltype").}

\item{cellrate}{The column name in \code{subcelldata} representing cell rates (default is "rate").}

\item{div_group}{The column name in \code{subcelldata} representing phenotype groups (default is "Phenotype").}
}
\value{
None
}
\description{
This function performs subset regression analysis for all cell types based on the provided data.
}
\examples{
cwas_regressiondraw(cwas.test.data)
}
