% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{detect_source_file}
\alias{detect_source_file}
\title{Detect the source file that called sparkle_app}
\usage{
detect_source_file()
}
\value{
The path to the source file, or NULL if not found
}
\description{
Detect the source file that called sparkle_app
}
\keyword{internal}
