% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-cache.R
\name{download_webr_package}
\alias{download_webr_package}
\title{Download a webR WASM package}
\usage{
download_webr_package(package_name, cache_dir, webr_version = "latest")
}
\arguments{
\item{package_name}{Name of the package to download}

\item{cache_dir}{Directory to store cached packages}

\item{webr_version}{WebR version to use (default: "latest")}
}
\value{
List with package name, version, and file paths
}
\description{
Downloads all 3 required files for a package: .tgz, .data, .js.metadata
}
\keyword{internal}
