\name{followButtons}
\alias{followButtons}
\title{Follow buttons}
\usage{
followButtons(small = TRUE, vertical = FALSE, facebook = NULL,
  twitter = NULL, linkedin = NULL, linkedin_company = NULL,
  google = NULL, youtube = NULL, flickr = NULL, vimeo = NULL,
  pinterest = NULL, instagram = NULL, foursquare = NULL, tumblr = NULL,
  rss = NULL)
}
\arguments{
  \item{small}{small size; TRUE or FALSE}

  \item{vertical}{vertical type; TRUE or FALSE}

  \item{facebook}{facebook user id}

  \item{twitter}{twitter user id}

  \item{linkedin}{linkedin user id}

  \item{linkedin_company}{linkedin company user id}

  \item{google}{google user id}

  \item{youtube}{user id}

  \item{flickr}{flickr user id}

  \item{vimeo}{vimeo user id}

  \item{pinterest}{pinterest user id}

  \item{instagram}{instagram user id}

  \item{foursquare}{foursquare user id}

  \item{tumblr}{tumblr user id}

  \item{rss}{rss feed}
}
\description{
...
}

