% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-cache.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get sparkle cache directory}
\usage{
get_cache_dir()
}
\value{
Path to cache directory
}
\description{
Returns the directory for cached webR WASM packages
}
\keyword{internal}
