\name{githubButton}
\alias{githubButton}
\title{Github buttons}
\usage{
  githubButton(user = "metagraf", repo = "sparkle",
    type = "watch", count = TRUE, small = TRUE)
}
\arguments{
  \item{user}{github user}

  \item{repo}{github repo}

  \item{type}{button type ("watch", "fork", "follow")}

  \item{count}{show count (TRUE/FALSE)}

  \item{small}{small size (TRUE/FALSE)}
}
\description{
  ...
}

