% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f12.1.Mediation_Analysis_kernal.R
\name{mediation_analysis_auto}
\alias{mediation_analysis_auto}
\title{Mediation Analysis Auto}
\usage{
mediation_analysis_auto(
  cwas.data,
  Best_model,
  X.cell = NULL,
  X.gene = NULL,
  Moderation.cell = NULL,
  Moderation.gene = NULL,
  nsim = 100,
  seed = 1,
  CI = "boot",
  fix_effect = NULL,
  random_effect = NULL
)
}
\arguments{
\item{cwas.data}{Data frame containing cell types, gene expression, and phenotype data.}

\item{Best_model}{List containing model selection information.}

\item{X.cell}{Character. Name of the independent cell type variable. Default is NULL.}

\item{X.gene}{Character. Name of the independent gene variable. Default is NULL.}

\item{Moderation.cell}{Character. Name of the mediating cell type variable. Default is NULL.}

\item{Moderation.gene}{Character. Name of the mediating gene variable. Default is NULL.}

\item{nsim}{Integer. Number of simulations. Default is 100.}

\item{seed}{Integer. Random seed for reproducibility. Default is 1.}

\item{CI}{Character. Method for confidence interval estimation. Default is "boot".}

\item{fix_effect}{Vector. Fixed effect covariates. Default is NULL.}

\item{random_effect}{Vector. Random effect covariates. Default is NULL.}
}
\value{
A list containing the results of the mediation analysis from the \code{bruceR::PROCESS} function.
}
\description{
This function performs mediation analysis using cell types or genes as independent and/or mediating variables. It processes the input data and leverages the \code{bruceR::PROCESS} function to conduct the analysis.
}
\details{
The function follows these main steps:
\enumerate{
\item \code{formula_process_string_vector}: Helper function to process formula strings.
\item \code{getfomula}: Helper function to extract formula information from \code{Best_model}.
\item Data preprocessing: Cleans and formats the data based on the specified independent and mediating variables.
\item Mediation analysis: Uses \code{bruceR::PROCESS} to perform mediation analysis and returns the results.
}

If \code{X.cell} is specified, the function will treat the cell type as the independent variable. If \code{X.gene} is specified, it will treat the gene as the independent variable. The mediating variable can be either a cell type or a gene, specified by \code{Moderation.cell} and \code{Moderation.gene} respectively.
}
\note{
Either \code{X.cell} or \code{X.gene} must be provided. If \code{X.gene} is used, \code{Moderation.cell} must be specified. The data frame \code{cwas.data} must contain a \code{Sample} column for merging operations.
}
\examples{
result <- mediation_analysis_auto(
  cwas.data = cwas.test.data,
  Best_model = cwas.test.data.single.model.best,
  X.cell = "MTF",
  Moderation.cell = "PIF",
  Moderation.gene = "TGFB1",
  nsim = 500,
  seed = 42,
  CI = "boot"
)

}
