% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f12.2.Mediation_forest_plot_v2.R
\name{Mediation_forest_plot}
\alias{Mediation_forest_plot}
\title{Generate a Mediation Forest Plot}
\usage{
Mediation_forest_plot(data)
}
\arguments{
\item{data}{A data frame containing the necessary information for plotting the mediation forest plot.}
}
\value{
A mediation forest plot.
}
\description{
This function generates a mediation forest plot to visualize mediation analysis results.
}
