% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f13.1.Moderation_Analysis_kernal.R
\name{moderation_analysis_auto}
\alias{moderation_analysis_auto}
\title{Moderation Analysis Function}
\usage{
moderation_analysis_auto(
  cwas.data,
  Best_model,
  X.cell = NULL,
  X.gene = NULL,
  Moderation.cell = NULL,
  Moderation.gene = NULL,
  nsim = 1000,
  seed = 1,
  CI = "boot"
)
}
\arguments{
\item{cwas.data}{A data frame containing CWAS data.}

\item{Best_model}{A list containing the best model chosen for the analysis.}

\item{X.cell}{A character string specifying the cell type to be used as the independent variable (IV). Default is NULL.}

\item{X.gene}{A character string specifying the gene to be used as the independent variable (IV). Default is NULL.}

\item{Moderation.cell}{A character string specifying the cell type to be used as the moderator. Default is NULL.}

\item{Moderation.gene}{A character string specifying the gene to be used as the moderator. Default is NULL.}

\item{nsim}{An integer specifying the number of simulations for bootstrapping. Default is 1000.}

\item{seed}{An integer specifying the seed for random number generation. Default is 1.}

\item{CI}{A character string specifying the confidence interval method. Default is "boot".}
}
\value{
A list containing the results of the moderation analysis.
}
\description{
This function performs moderation analysis on CWAS data. It supports different types of moderation including cell-cell, gene-cell, and gene-gene mediation effects.
}
