% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f8.4.OR_plot.R
\name{OR_plot}
\alias{OR_plot}
\title{OR Plot Function}
\usage{
OR_plot(
  data,
  show_variable = c("Celltype", "Model", "OR(95\%CI)", "P-value", "P-value adjusted"),
  plot.position = 4,
  Tablename = "Tablename"
)
}
\arguments{
\item{data}{A list containing required data for generating the forest plot.}

\item{show_variable}{A character vector specifying the variables to display in the forest plot.
Default is c("Celltype", "Model", "OR(95\%CI)", "P-value", "P-value adjusted").}

\item{plot.position}{An integer specifying the position to insert an empty space in the \code{show_variable} vector.
Default is 4, which corresponds to inserting an empty space after "P-value".}

\item{Tablename}{A character string specifying the table name to display in the forest plot header.
Default is "Tablename".}
}
\value{
A forest plot visualizing odds ratios and confidence intervals based on the provided data.
}
\description{
This function generates a forest plot to visualize odds ratios (OR) and confidence intervals (CI) from provided data.
}
\examples{
# Example usage:
OR_plot(data = cwas.test.data.single.model.best)

}
\keyword{plot}
