% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{print.sparkle_callback}
\alias{print.sparkle_callback}
\title{Print method for sparkle callbacks}
\usage{
\method{print}{sparkle_callback}(x, ...)
}
\arguments{
\item{x}{A sparkle_callback object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for sparkle callbacks
}
