% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{print.sparkle_element}
\alias{print.sparkle_element}
\title{Print method for sparkle elements}
\usage{
\method{print}{sparkle_element}(x, ...)
}
\arguments{
\item{x}{A sparkle_element object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for sparkle elements
}
