% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{serve_static_file}
\alias{serve_static_file}
\title{Serve a static file with appropriate headers}
\usage{
serve_static_file(file_path, extra_headers = list())
}
\arguments{
\item{file_path}{Path to file to serve}

\item{extra_headers}{Additional headers to include}
}
\value{
httpuv response list
}
\description{
Serve a static file with appropriate headers
}
\keyword{internal}
