% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.rds.R
\name{sf.rds}
\alias{sf.rds}
\title{Save variable as .rds file}
\usage{
sf.rds(variable, file = "viarablename")
}
\arguments{
\item{variable}{The R object to be saved.}

\item{file}{Character string specifying the directory path where the file will be saved.}
}
\value{
A message indicating the file name and directory where the .rds file has been saved.
}
\description{
This function saves a variable as an .rds file in a specified directory.
}
