% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled.R
\name{snake_to_kebab}
\alias{snake_to_kebab}
\title{Convert snake_case to kebab-case for CSS properties}
\usage{
snake_to_kebab(str)
}
\arguments{
\item{str}{Character string in snake_case}
}
\value{
Character string in kebab-case
}
\description{
Convert snake_case to kebab-case for CSS properties
}
\examples{
\dontrun{
snake_to_kebab("background_color")  # "background-color"
snake_to_kebab("font_size")         # "font-size"
}
}
\keyword{internal}
