% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{sparkle_get_state}
\alias{sparkle_get_state}
\title{React Hooks for Sparkle Components}
\usage{
sparkle_get_state(index)
}
\arguments{
\item{index}{The index of the state variable}
}
\value{
The current value of the state
}
\description{
These functions provide access to React's hooks system from R code.
They manage state that bridges to React via the JavaScript runtime.
Global state accessor functions
}
\details{
Get the current value of a state variable by its index.
This is used internally by callbacks to access state.
}
