% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{sparkle_set_state}
\alias{sparkle_set_state}
\title{Set the value of a state variable (internal)}
\usage{
sparkle_set_state(index, value)
}
\arguments{
\item{index}{The index of the state variable}

\item{value}{The new value}
}
\value{
A list signaling the state update to JavaScript
}
\description{
Update a state variable and signal to JavaScript that state has changed.
This triggers a re-render of the component.
}
\keyword{internal}
