% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled.R
\name{styled-components}
\alias{styled-components}
\title{Styled Components for Sparkle}
\description{
Create styled components with CSS-in-R. These functions return factory functions
that create elements with automatically generated CSS classes.
}
