% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{validate_app_function}
\alias{validate_app_function}
\title{Validate that App function is defined in code}
\usage{
validate_app_function(serialized_code)
}
\arguments{
\item{serialized_code}{String containing R code}
}
\value{
TRUE invisibly if valid, stops with error otherwise
}
\description{
Validate that App function is defined in code
}
\keyword{internal}
