% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{wrap_fn}
\alias{wrap_fn}
\title{Event Callback Wrapper}
\usage{
wrap_fn(fn)
}
\arguments{
\item{fn}{An R function to wrap. Can be anonymous: \code{\\() expr}}
}
\value{
A wrapped function object that can be passed to event props
}
\description{
Wraps R functions for use as event handlers in Sparkle components.
The wrapped function will be serialized and executed asynchronously
when the event fires.
Wrap a function for use as an event callback
}
\details{
This function wraps an R function so it can be used as an event handler
(e.g., on_click, on_change). The function will be serialized and executed
asynchronously in webR when the event fires.
}
\examples{
\dontrun{
tags$button(
  "Click me",
  on_click = wrap_fn(\() {
    print("Button clicked!")
  })
)
}
}
