\name{followButtons}
\alias{followButtons}
\title{Follow buttons}
\usage{
  followButtons(small = TRUE, vertical = FALSE,
    facebook = NULL, twitter = NULL, linkedin = NULL,
    linkedin_company = NULL, google = NULL, youtube = NULL,
    flickr = NULL, vimeo = NULL, pinterest = NULL,
    instagram = NULL, foursquare = NULL, tumblr = NULL,
    rss = NULL)
}
\arguments{
  \item{style}{style type (1-7)}
}
\description{
  ...
}

