% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\docType{data}
\name{tags}
\alias{tags}
\title{HTML Tag Builders}
\format{
An object of class \code{list} of length 9.
}
\usage{
tags
}
\description{
Create virtual DOM descriptions for HTML elements. These functions return
list structures that the Sparkle JavaScript bridge converts into React elements.
}
\keyword{datasets}
