% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{use_state}
\alias{use_state}
\title{Use State Hook}
\usage{
use_state(initial_value)
}
\arguments{
\item{initial_value}{The initial value for the state variable}
}
\value{
A list with two elements:
  \item{index}{The index of this state variable (for use in callbacks)}
  \item{value}{The current state value}
}
\description{
Creates a state variable similar to React's useState hook.
The state is managed by React in JavaScript, and this function
provides an R interface to it.
}
\examples{
\dontrun{
Counter <- function() {
  count <- use_state(0)
  tags$div(
    tags$h1(paste("Count:", count$value)),
    tags$button(
      "Increment",
      on_click = wrap_fn(function() {
        current <- sparkle_get_state(count$index)
        sparkle_set_state(count$index, current + 1)
      })
    )
  )
}
}
}
